Post functionalities added

+++ Mockapp/views/posts.py
@@ -1,4 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
 from ..models import Post
 from .forms import PostForm
 
@@ -8,28 +10,43 @@ def post_index(request):
 
 def post_detail(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    return render(request, 'posts/detail.html', {'post': post})
+    return render(request, 'posts/show.html', {'post': post})
 
+@login_required
 def new_post(request):
     if request.method == 'POST':
         form = PostForm(request.POST)
         print(form)
         if form.is_valid():
+            post = form.save(commit=False)
+            post.user = request.user
+            post.save()
 
-            form.save()
-            return redirect('post_list')
+            return redirect('post_index')
     else:
         form = PostForm()
 
     return render(request, 'posts/new.html', {'form': form})
 
+@login_required
 def post_edit(request, pk):
-    # Implement your edit view logic here
     post = get_object_or_404(Post, pk=pk)
+    if post.user != request.user:
+        return HttpResponseForbidden("You are not allowed to edit this post.")
+    
+    if request.method == 'POST':
+        form = PostForm(request.POST, instance=post)
+        if form.is_valid():
+            form.save()
+            return redirect('post_index')
+    else:
+        form = PostForm(instance=post)
     return render(request, 'posts/edit.html', {'post': post})
 
+@login_required
 def post_delete(request, pk):
-    # Implement your delete view logic here
     post = get_object_or_404(Post, pk=pk)
+    if post.user != request.user:
+        return HttpResponseForbidden("You are not allowed to delete this post.")
     post.delete()
-    return redirect('post_list')
+    return redirect('post_index')
